Automatically make the models available to gazebo_ros
By adding gazebo_model_path to the exports of this package, the models
directory will be automatically available if you are using
`gazebo.launch.py`.  This means that you don't need to set
GAZEBO_MODEL_PATH this in your .bashrc or in other places.

This was implemented in the ROS1 branch but the change was
somehow forgotten.